restructured visualisation, added init files for everything
