+++ peepingtom/__init__.py
@@ -1,1 +1,2 @@
-
+from ._io import star_to_crates, images_to_crates
+from .visualisation import Peeper, ParticlePeeper
