+++ peepingtom/core/__init__.py
@@ -1,3 +1,4 @@
-from .datacrate import DataCrate
-from .datablock import DataBlock, PointBlock, LineBlock, OrientationBlock, ImageBlock
-from .groupblock import GroupBlock, ParticleBlock
+from .base import DataBlock, GroupBlock, DataCrate
+
+from .datablocks import DataBlock, PointBlock, LineBlock, OrientationBlock, ImageBlock, \
+    LineBlock, SphereBlock, PropertyBlock, ParticleBlock
