+++ peepingtom/visualisation/__init__.py
@@ -1,1 +1,4 @@
+from .base import Depictor
+from .depictors import ImageDepictor, ParticleDepictor
 from .peeper import Peeper
+from .peepers import ParticlePeeper
