+++ peepingtom/visualisation/depictors/__init__.py
@@ -1,0 +1,2 @@
+from .images import ImageDepictor
+from .particles import ParticleDepictor
