+++ peepingtom/visualisation/gui/__init__.py
@@ -1,0 +1,1 @@
+from .gui import make_property_sliders, make_categorical_checkboxes
