+++ peepingtom/visualisation/peepers/__init__.py
@@ -1,0 +1,1 @@
+from .particle_peeper import ParticlePeeper
